requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
django-hashid-field==3.3.7
djangorestframework==3.14.0
hashids==1.3.1
importlib-metadata==5.1.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
zipp==3.11.0
django-cors-headers==3.13.0

gunicorn==20.1.0

Pillow==9.1.1

psycopg

psycopg2-binary

python-dateutil==2.8.2

whitenoise==6.1.0
